Refactor common TE BYOC code into separate file
